Add parameter bounds to function type data structure.
This changes adds a variable-sized array for parameter bounds to the function
type data structure.  clang has an ExtProtoInfo structure that holds additional
information about prototypes that we extend also. For now, we just test
that empty arrays work.  We don't actually exercise populating the
parameter bounds in ExtProtoInfo.  We still have to update AST importing
and exporting to save/restore the parameter bounds.

Testing:
- Passes Checked C regression tests.
- Passes clang regression tests.