Add way to represent parameters positionally.
For bounds expressions in function types, we will represent references
to parameter variables positionally, not using the names.  This allows to
disregard the parameter names in the type.  This change adds an expression
for representing positional parameters.  It also includes code for
canonicalizing bounds expressions to use the positional parameters.
That code is currently not tested.

Testing:
- Passes Checked C regression tests.
- Passes clang regression tests.