Finish adding bounds expressions for parameters in function types.
This change finishes adding bounds expressions for function parameter
types. It introduces PositionalParameterExpr, which represents
a parameter by its index in a parameter list, not by its name.
This requires a fair amount of boiler-plate updates throughout
the code.

It changes the code for building function types to canonicalize
bounds expressions before adding them to the array of parameter
bounds.

It adds tests of AST dumping that check that bounds expressions for
parameters have been rewritten to use positional parameters.

It updates AST reading and writing classese to include the parameter
bounds array, if there is one.

It updates the design documentation to describe the IR choice
we have made for Checked C and why, including the choice
to use PositionalParameterExpr.

Testing:
- Passes Checked C regression tests.
- Passes existing clang regression tests.