Add return bounds information to function types.
We add a member to FunctionType to hold the return bounds expression.  The
return bounds expression is abstracted before it is added to the function
type.  References to parameters are replaced with PositionalParameterExpr's.

We could save space for FunctionType objects that do not have return bounds
by storing the return bounds as part of the dynamically-allocated array for
parameter bounds.  We do not do that for now. There's no pressing need to
save space and it would make the implementaiton a little more complicated.

Testing:
- Adding new AST dumping tests that check that return bounds information has
  been added to function types.
- Passes existing Checked C regression tests.
- Passes existing clang regression tests.