Check bounds declarations for function pointer return types. (#90)
This change adds checking of bounds declarations for function pointer return types.  This addresses GitHub issue #89.   The checking was being done as part of constructing a function declaration.  However, function declarations are not constructed for variables with function pointer types.  Move the checking from construction of function declarations to construction of function types.

To support this, refactor the code for checking bounds declarations so that there are now separate methods for (1) checking that a bounds expression is valid for a type (2) checking that a bounds declaration is valid for a declaration.  Checking for function return types invokes the first method.   The existing checking for variable declarations invokes the first method.

As part of the refactoring, check that bounds-safe interface types are not declared for local variables. This addresses Github issue #77.

Revamp the error messages for return types that have bounds declarations and bounds-safe interface type annotations.  There may not be a named entity to refer to in the error messages, so remove mention of that and just point at the problematic bounds expression or annotation type in the error
message.

It may be useful to make the error messages for variable declarations with bounds be similar to the error messages for return types.  Put that off for now to avoid including an unrelated change with this change.

Testing:
- There will be a separate commit updating the Checked C regression tests in the Checked C repo.   That commit will:
-- Fix errors found in existing tests because of the additional checking
-- Add additional tests of function pointer parameters and returns with bounds declarations and bounds-safe interfaces.
-- Update existing error messages involving bounds declarations and return types
-- Add tests making sure that bounds-safe interface types are not allowed for local variables.
-- Add tests for unnamed parameters, to make sure that compilers handle  this.  The clang error messages are confusing when the parameter has no name.  We can fix this as part of updating the error messages for variable declarations with bounds.
- Update Checked C clang tests.  One test checked for two error messages, but there is only one now.  It is covered by existing tests, so delete it.   Another test was declaring an interface type for a local variable, which is  not allowed, so delete that case.
- Passes existing clang regression tests.