Check redeclarations of functions with bounds declarations. (#91)
Now that bounds information is incorporated into function types, use it when determining function type compatibility.  This allows the existing machinery for checking redeclarations of functions to check redeclarations of functions with bounds declarations too.   It checks that the redeclaration has equivalent bounds declarations to the original declaration.   It also checks redeclarations of functions with bounds-safe interfaces.

For now, require that the bounds expressions for parameters or returns be identical syntactically for function types, modulo parameter names.  We will eventually loosen this requirement to be bounds expressions that  are semantically equivalent (for example, canonicalized bounds expressions).   Bounds-safe interfaces are treated specially: a parameter or return with unchecked pointer type and no bounds expression is compatible with one with a bounds expression or an interface type.   If both parameters or returns have bounds expressions or interface types, the expressions or types must be identical.

To determine whether bounds expressions are equivalent, re-use the existing clang "profile" code for determining whether two expressions are equivalent.

The main implementation change is to update mergeFunctionTypes in ASTContext.cpp to include bounds information.  This allows the replacement or deletion of special-case code that worked around the lack of bounds information in function type.
- Code is deleted in Sema::DiagnoseCheckedCFunctionCompatibility, Sema::CheckedCFunctionDeclCompatibility, and GetFullTypeForDeclarator in SemaType.cpp.   For GetFullTypeForDeclarator, the new code in isNotAllowedForNoPrototypeFunction subsumes this check.
- Code is updated in ASTContext::isNotAllowedForNoPrototypeFunction. 

A function can be declared via typedef. In this case, clang synthesizes parameters from the function prototype in Sema::ActOFunctionDeclarator.  Extend this code to  synthesize bounds information from the prototype too.  Bounds expressions are abstracted with respect to parameter names in function types using positional indicates.  The bounds expressions have to be concretized with respect to the newly built parameters.

Currently the error messages say "type mismatch" when bounds do not match.    There is a separate work item for improving the error messages (GitHub #83).

Testing:
- A separate commit to the Checked C repo will add tests of redeclarations of functions.  It will also add header files that add Checked C bounds-safe interfaces to functions in the C standard library.
- Passes existing Checked C tests.
- Passing existing clang regression tests.