Restrict variable references in bounds expressions in function types.
We have chosen a representation of bounds expressions for function types
that only supports references to global variables or parameters declared
in the function declarator for a function type.  Typechecking will not work
properly for the cases of references to variables in enclosing function
declarators.  The uses of local variables in function types will always
lead to errors.   There is no way to create a non-null function pointer with a
type that involves a local variable.

Add checks that detect the unsupported usage of variables in bounds expressions
for function types.  Also add clang-specific tests that detect this.

For what is worth, we could probably support references to parameters
in enclosing function declarators pretty straightforwardly.  Local variables
would be much more challenging.  Typechecking would involve reasoning
about the values of those local variables.  We'd be into the
dependently-typed case that isn't supported by existing type checkers
for C, even for variable-length arrays.  We would need to extend typechecking
of function types with bounds expressions with a context. In any case, both
cases would require some thought, so prevent their use right now.

Testing:
- Passes existing Checked C regression tests.
- Passes clang test suite.