Error messages for mismatched bounds declarations.
Add better error messages for mismatched bounds declaration when functions are
redeclared.  The current error message is just "type mismatch", which is not
helpful.

With this change,
- Errors that involve only bounds declarations are diagnosed as mismatched
  bounds errors.
- Type compatibility errors involving prior no-prototype and prototype
  declarations of the same function continue to be diagnosed
  specially.
- If a function declaration has both bounds declarations mismatches
  and a type mismatch when bounds are ignored, there will be error messages
  for both.

When an error message is emitted for mismatched bounds declarations, also
emit a note that points at the most specific context from the prior declaration
that  is helpful for understanding the error.  By being specific, this makes
it easier for programmers to understand the error.
- If the prior declaration has a bounds expression, point at that.
- If not, use the prior declaration.  This will be the parameter declaration
  for parameter bounds, which is still pretty specific. It will be the entire
  function declaration for return bounds.

Identifying prior bounds expressions requires some special care for
unchecked pointer types.  A declaration of a function with a bounds-safe
interface for a parameter or return with unchecked pointer type is compatible
with a declaration with no bounds-safe interface.  The prior declaration
may not actually have a bounds expression for a parameter or return. An
even earlier definition may have the bounds expression, so we need to
search for that.

We identify errors that only involve bounds declarations by checking
whether type merging succeeds when bounds declarations are ignored.  If it
does, this is a "bounds-only" error.   To implement this, parameterize type
merging by an additional parameter that controls whether bounds information is
ignored.

Testing:
- Passes Checked C regression tests, once the error messages for regression
  tests of bounds mismatches have been updated.
- Still to do: test the cases where bounds declarations are omitted are
  added.