Add checking of redeclarations of variables with bounds. (#96)
This change adds checking of redeclarations of variables with bounds, ensuring that the redeclarations follow the Checked C rules.  This completes the work for issue #30.  For now, when bounds are required to match, we require them to be identical syntactically.  This will be generalized later.

The checking is done when bounds are attached to variable declarations, not during the merging of declarations.  Bounds are attached to variable declarations after declarators have been processed (the bounds may refer to the variable being declared, so the variable declaration needs to be built before we build the bounds expression).  This means bounds can't be checked during merging of declarations, which operates on just the declarators.
- ActOnBoundsDecl does the checking for conflicting bounds expressions on variable declarations.
- Generalize the existing code for checking for conflicting bounds expressions on declaration to handle non-parameter variables.  Also generalize the check for variables with unchecked types to include unchecked arrays.   Declarations of variables with unchecked pointer and unchecked array types and bounds-safe interfaces are compatible with the declarations that omit the bounds-safe interfaces.
- Rework the existing error messages for variable redeclarations to use the clang select mechanism for error messages.   This lets us use one diagnostic id in place of several diagnostic ids and simplify the code.

I discovered an error in the parsing of bounds expressions for function declarators.  I believe it was possible to write a bounds expression after a function declarator, which would have triggered an internal compiler assert:  `array_ptr<int> f(int len) : count(len) : count(5)`, for example.   The fix is to not try to parse a bounds expression after a function declarator.  The parsing of the function declarator already handled the bounds expression.

Testing:
- Added tests for redeclarations of variables to the Checked C regression tests.  This will be committed separately to the Checked C repo.
- Passes existing Checked C tests.
- Passes clang regression test suite.