Actually traverse expressions.
- Fix the code so that it actually traverses expressions.
- Fix assert for NullaryBoundsExpression to allow Any as a
  bounds kind.
- Assignment expressions are now traversed and bounds
are inferred for right-hand sides where the targets of
the assignments have bounds requirements.

Right now we are only checking that the right-hand side
of the expression is an expression form that that might have
a bounds expression.  Bounds(any) is being inferred for any
expression that might have a bounds.  We'll add code that
infers that actual bounds case-by-case.

 Testing:
 - Checked C regression tests pass.