Walk IR for static checking of bounds declarations. (#110)
Add initial code for walking expressions and declarations for static checking of bounds declarations.  The checking code will be invoked by this walking code.

To test the walking code, check some simple properties of bounds for assignments and variables declarations.    If the left-hand side of an assignment has a bounds requirement, check that the right-hand side of the assignment has the right form to have an inferred bounds expression.   This is done by calling a stubbed out inference method that returns bounds(any) for expression forms that might have a bounds expression.  Do a similar check for initializers for variables declarations with bounds.  Also check that automatic variables that have bounds declarations or _Ptr type have initializers.

Add the 'Any` bounds to the IR for bounds expressions.

Testing:
- There will be a separate pull request for the GitHub Checked C repo that updates tests where initializers were missing.    The pull request will also add tests of error messages for missing initializers.
- Passes existing clang regression tests.