Add Checked C's _Dynamic_check including Code Generation
We implement it as a clang builtin, which means we get compiler hooks for code generation and semantic analysis without having to extend the lexer/parser. 

For the control flow I have followed the design in docs/checked/Dynamic-check.md

The largest omission at the moment is a set of comprehensive tests, and semantic analysis to make sure it's only called on non-modifying expressions. 

References #114.
Closes #109.