Adds Non-Modifying Expression check to Dynamic Checks (#122)
Though we have not yet updated the specification, our intention is to require explicit `dynamic_check`s contain non--modifying expressions, in part so we can run them as and when we want to, without fear of side effects. This also partly conveys the expectation that these checks should be lightweight.

I have implemented this as a new `RecursiveASTVisitor` in `SemaBounds.cpp`. If this visitor finishes early, it shows we found something not allowed in a non-modifying expression. The error message tries to clarify what part of the expression exactly we don't like.