Infer bounds for address-of expressions. (#126)
This change adds inference of bounds of address-of expressions.   address-of expression converts an expression that is an lvalue to a pointer.   The bounds of the address-of expression are the bounds of the lvalue.

This change covers address expressions of the form &v, where v is a variable, &*e1, or &e1[e2]. The change adds inference of lvalue bounds for non-array variables, pointer dereference expressions, and array-subscript expressions.
 
Add a comment for ExpandToRange, as requested by  a code review for a prior pull request.

Testing:
- Add feature tests of bound inferences for the various kinds of address-of expressions.
- Passes existing Checked C tests.
- Passes existing clang tests.