Explicit dynamic check warning flag (#132)
Adds a command-line switch for the Checked C warnings. The switch is called `-Wcheckedc`. I think passing `-Wno-checkedc` will disable the warnings, as the example clang pragma also does. 

One of the Clang tests verifies that if we add more warnings, then we should add a command-line switch so users can disable this warning.