Explicit dynamic check statistics (#135)
In order to prepare for adding a lot more code generation for dynamic checks, I have moved some code around and added statistics. 

Statistics of code generation can be viewed with the `-print-stats` argument to `clang -cc1`. 

I have also moved the warning about failing dynamic checks into semantic analysis instead of code generation, because this seems like a much better place for it. It used to only be added when you asked `clang -cc1` to do codegen (via `-emit-llvm` or similar), but now it's run with all the other semantic analysis, so I had to update some tests which now produced warnings.