Warn on invalid use of Checked C extension flag. (#137)
The Checked C clang compiler issues an error message when the
`-fcheckedc-extension` is used with an invalid language target (such as C++
or Objective C).   This change modifies to issue a warning message
and ignore the flag instead.  This allows the -fcheckedc-extension flag to be
added as an additional compiler flag in build systems that compile C and C++
without breaking builds.

I hit this issue using the LNT --cflags option to pass the -fcheckedc-extension
flag to the compiiler when running LLVM nightly tests.  The flag was passed to
compilations of C++  programs, which caused them to fail to compile.

This change updates the driver tests that use the -fcheckedc-extension
flag while compiling languages other than C to reflect the new behavior.

Testing:
- The nightly tests in the original branch of
checkedc-llvm-test-suite pass when the -fcheckedc-extension
flag is set (2495 tests pass).  Of course, none of these tests use the
Checked C extension, so this just verifies that we haven't broken any
of the tests with the modifications for Checked C.
- Ran the driver tests using LIT.