Check additional expressions where bounds are required.
When memory is read or written through a pointer that is derived
from an _Array_ptr, the pointer must have bounds.  In addition, the
source expression of a  cast to _Ptr type must have bounds.

In C, a ead through a pointer is done by an lvalue-to-rvalue cast and write
through a pointer is done by assignment operation.
- Add a function LValueDerivedFromArrayPtr that determines if an lvalue is
  derived from an array_ptr value.
- Use this function to check at an assignment whether a bounds check is needed
  before writing to the LHS of the assignment.  If it is, check that bounds
  can be inferred for the LHS lvalue.
- Also use this function at lvalue-to-rvalue casts to check if a bounds
  check is needed before reading memory.  If it is, check that bounds can
  be inferred for the lvalue.
- Add a check at casts to _Ptr type that the source expression has bounds.
- Add code to dump bounds for these new cases.

Clang represents parameters with array types with pointer types.  For
parameters with checked array types, we need to implicitly add a bounds
expression as part of the retyping, so that bounds checking works out.

Add a method that creates a count expression from constant-dimension
array type.  Refactor existing code in SemaBounds that already did this
into a separate methodn.

Testing:
- Existing Checked C and clang regression tests pass.  The additional
  checking found places in Checked C tests where bounds declarations were
  missing or illegal assignments from unchecked pointers to checked pointers
  were occurring.
- Add a test to ast-dump-bounds that tests that checked array parameters
  have a bounds declaration added.
- Additional tests are still needed before this is ready to commit to
  the Github repo.