Avoid converting to _Ptr type for pointer comparison.
Implicit conversions to _Ptr type are only allowed if the source
operand has bounds suitable for _Ptr type.   This collides with
existing logic for allowing pointer comparison between
incompatible pointer types.  The existing code inserts a conversion
to the type of one of the operands. Sometimes it picks a target
type of _Ptr.  This leads to a bounds checking failure if the
operand does not have suitable bounds.

This change modifies the code to choose a non-_Ptr type as the
target type.  If both types are _Ptr types, both operands are
implicitly converted to an unchecked type.

Fro now, we always trust implicit conversions from checked types
to unchecked types.  These are only inserted if the value with
unchecked type is not used to access memory.  We will want to
revisit this when we add support to the IR for assume_cast
operations.