Address code review feedback for bounds inference and failing test. (#161)
Code review pointed out that the code is inconsistent about invariants for
bounds inference: sometimes the code returns an invalid bounds expression,
sometimes it returns a bounds of None, and sometimes it returns a nullptr.
This can be confusing. We want the IR invariant that:
1. If an expression needs to have bounds checking, the inferred bounds property
   should either be a valid bounds expression or an invalid bounds expression
   (if the expression has no bounds when bounds are expected)
2. If an expression does need bounds checking, it has no inferred bounds
   expression.

Clean up the code to make it clearer that this IR invariant is being
established by bounds inference:
- Have a separate by-reference variable that indicates whether bounds checking
  is expected or not, instead of using a nullptr to indicate no bounds checking
  is expected.
- When a bounds expression is expected, return a valid bounds expression.  If
  the bounds expression is None, convert it to an invalid bounds expression.

Fix the failing inferred_bounds\basic.c test.  There are tests where bounds
inference errors are expected to be generated because an expression has
bounds of None when valid are expected.   Replace Bounds(None) with
Bounds(Invalid) for the AST dumps for those cases.  This addresses issue #188.