Avoid converting to ptr type for pointer comparison. (#162)
Implicit conversion to a `ptr` type is only allowed if the source operand has bounds suitable for the `ptr` type.  This collides with existing logic for allowing pointer comparison between incompatible pointer types, a language extension supported by clang.  The existing logic inserts a conversion to the type of one of the operands.  Sometimes it picks a target type of `ptr`.  This leads to a bounds checking failure if the operand does not have suitable bounds.

This change modifies the code to choose a non-`ptr` type as the target type. If both types are `ptr` types, both operands are converted implicitly  to an unchecked type.  This addresses issue #155 and allows the error message for bounds inference for casts to `ptr` type to be enabled.

For now, we always trust implicit conversions from checked types to unchecked types.  These are only inserted if the value with unchecked type is not used to access memory.  We will want to revisit this when we add support to the IR for `assume_cast` operations.