Check operands for increment and decrement operations.
The operands for increment and decrement operators  must have bounds
if the operand is an array_ptr dereference.

Now that the IR includes inferred bounds expressions, when the dumpBounds
flag is set, just dump the IR for the increment/decrement expression after
bounds have been inferred.   There is no need for to pretty-print the
results of the bounds inference.

Testing:
- Checked C tests pass.
- clang regression tests for Checked C pass.
- Add test for inference of bounds expressions for increment and
  decrement operations.