Remove duplicate information printed under DumpBounds.
AST expression dumps now include inferred bounds information.  This change
removes duplicate bounds information for casts and member reference expressions
printed under dumpBounds.  We don't need to include the inferred information
separately because it is part of the dump of the expression.  We can
also remove the specialized functions for dumping the results of bounds
inference for these expressions.

Testing:
- Checked C regression tests pass.
- Checked C clang regression tests pass.