Dynamic bounds check array subscript (#163)
Closes #145 

Code generation in Clang is really complicated, with 5 different code generators (4 for each of the evaluation kinds, and then an lvalue code generator) each recursively calling each other through visitors and the like. The aim here has been to have as little overall effect on the code as possible, while still inserting the checks. This is why we save BoundsExprs from where they're inferred for until later in the CodeGenFunction instance, because we need them only after we have Emitted the value that calculates the pointer address.

I think doing it this way preserves an invariant I believe clang has - that we only call Emit on an expression once, which should translate into making sure we don't evaluate any side-effects more than they would be normally, for instance in checking the expression `e[i++]`